pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deductoscope"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "colorama>=0.4.6",
    "requests>=2.31.0",
    "dnspython>=2.4.0",
    "python-whois>=0.8.0",
    "beautifulsoup4>=4.12.0"
]

[project.scripts]
deductoscope = "deductoscope.cli:main"